+++ R/datatables.R
@@ -88,10 +88,25 @@
 #'   column, or \code{"all"} to edit all cells on the current page of the table.
 #'   In all modes, start editing by doubleclicking on a cell. This argument can
 #'   also be a list of the form \code{list(target = TARGET, disable =
-#'   list(columns = INDICES))}, where \code{TARGET} can be \code{cell},
-#'   \code{row}, \code{column}, or \code{all}, and \code{INDICES} is an integer
-#'   vector of column indices. Use the list form if you want to disable editing
-#'   certain columns.
+#'   list(columns = INDICES))}, where \code{TARGET} can be \code{"cell"},
+#'   \code{"row"}, \code{"column"}, or \code{"all"}, and \code{INDICES} is an
+#'   integer vector of column indices. Use the list form if you want to disable
+#'   editing certain columns. You can also restrict the editing to accept only
+#'   numbers by setting this argument to a list of the form \code{list(target =
+#'   TARGET, numeric = INDICES)} where \code{INDICES} can be the vector of the
+#'   indices of the columns for which you want to restrict the editing to
+#'   numbers or \code{"all"} to restrict the editing to numbers for all columns.
+#'   If you don't set \code{numeric}, then the editing is restricted to numbers
+#'   for all numeric columns; set \code{numeric = "none"} to disable this
+#'   behavior. Finally, you can also edit the cells in text areas, which are
+#'   useful for large contents. For that, set the \code{editable} argument to a
+#'   list of the form \code{list(target = TARGET, area = INDICES)} where
+#'   \code{INDICES} can be the vector of the indices of the columns for which
+#'   you want the text areas, or \code{"all"} if you want the text areas for
+#'   all columns. Of course, you can request the numeric editing for some
+#'   columns and the text areas for some other columns by setting
+#'   \code{editable} to a list of the form \code{list(target = TARGET, numeric
+#'   = INDICES1, area = INDICES2)}.
 #' @details \code{selection}:
 #'   \enumerate{
 #'     \item The argument could be a scalar string, which means the selection
@@ -270,8 +285,13 @@ datatable = function(
   params$caption = captionString(caption)
 
   if (isTRUE(editable)) editable = 'cell'
-  if (is.character(editable)) editable = list(target = editable, disable = list(columns = NULL))
-  if (is.list(editable)) params$editable = editable
+  if (is.character(editable))
+    editable = list(target = editable, disable = list(columns = NULL))
+  if (is.list(editable)) {
+    editable$numeric = makeEditableNumericField(editable$numeric, data, rn)
+    editable$area = makeEditableAreaField(editable$area, data, rn)
+    params$editable = editable
+  }
 
   if (!identical(class(callback), class(JS(''))))
     stop("The 'callback' argument only accept a value returned from JS()")
@@ -362,6 +382,34 @@ datatable = function(
 
       instance
     }
+  )
+}
+
+makeEditableNumericField = function(x, data, rn) {
+  as.list(
+    if (is.null(x))
+      which(unname(vapply(data, is.numeric, logical(1L)))) - 1L
+    else if (identical(x, 'none'))
+      NULL
+    else if (identical(x, 'all'))
+      seq_along(data) - 1L
+    else if (is.null(rn))
+      x - 1
+    else
+      x
+  )
+}
+
+makeEditableAreaField = function(x, data, rn) {
+  as.list(
+    if (is.null(x))
+      NULL
+    else if (identical(x, 'all'))
+      seq_along(data) - 1L
+    else if (is.null(rn))
+      x - 1
+    else
+      x
   )
 }